requirements.txt
django-heroku
gunicorn
djangorestframework~=3.13.1
pytz
Django~=4.0.1